==> R/01_load.qmd <==
---
title: "01_load"
author: Antoine Andréoletti (s242830), Olivier Gaufrès (s243252), Amy Surry (s201902), Lea Skytthe (s203531), Trine Søgaard (s204655)
format:
  html:
    embed-resources: true
editor: visual
---

# Uploading data

## Data

The data we are using in this exercise comes from here: <https://ftp.ebi.ac.uk/biostudies/fire/E-MTAB-/807/E-MTAB-12807/Files/>.

## Load Libraries

```{r}
#| label: 01_load-libraries
#| message: False
#| warning: False

rm(list = ls())

library("tidyverse")
library("httr")
library("XML")
library("stringr")
```

## Creating data folder

```{r}
#| label: 01_creating-data-folder

data_dir <- "../data"
results_dir <- "../results"
raw_data_dir <- "../data/_raw/"
count_data_dir <- "../data/_raw/count_data"
metadata_dir <- "../data/_raw/metadata"

if( !dir.exists(data_dir) ){
  dir.create(path = data_dir)
}

if( !dir.exists(results_dir) ){
  dir.create(path = results_dir)
}

if( !dir.exists(raw_data_dir) ){
  dir.create(path = raw_data_dir)
}

if( !dir.exists(count_data_dir) ){
  dir.create(path = count_data_dir)
}

if( !dir.exists(metadata_dir) ){
  dir.create(path = metadata_dir)
}
```

## Upload data

```{r}
#| label: 01_upload-data
#| message: False

base_url <- "https://ftp.ebi.ac.uk/biostudies/fire/E-MTAB-/807/E-MTAB-12807/Files/"

# Make an HTTP GET request
response <- GET(base_url)

# Parse the HTML content
html_doc <- htmlParse(content(response, as = "text"), asText = TRUE)

# Extract file links
file_links <- xpathSApply(html_doc, "//a/@href")
file_links <- data.frame(files = file_links)

count_files <- file_links |> 
  filter(str_ends(files, ".sf")) |> 
  pull(files)

metadata_files <- file_links |> 
  filter(str_ends(files, ".txt")) |> 
  pull(files)

# Importing each file and downloading it in the right folder
base <- "../data/_raw"

# Downloading count data files
for (count_file in count_files) {
  new_file_name <- str_replace(count_file, "_quant.sf", "_quant.csv")
  new_file_path <- paste(base, "/count_data/", new_file_name,
                         sep = "")
  if (file.exists(new_file_path)) {
    next
  }
  file_path <- paste(base_url, count_file, 
                     sep = "")
  file <- read_tsv(file = file_path)
  write_csv(file, file = new_file_path)
}

# Downloading metadata files
for (metadata_file in metadata_files) {
  new_file_name <- str_replace(metadata_file, ".txt", ".csv")
  new_file_path <- paste(base, "/metadata/", new_file_name,
                         sep = "")
  if (file.exists(new_file_path)) {
    next
  }
  file_path <- paste(base_url, metadata_file, 
                     sep = "")
  file <- read_tsv(file = file_path)
  write_csv(file, file = new_file_path)
}
```
