==> R/04_describe.qmd <==
---
title: "04_describe"
author: Antoine Andréoletti (s242830), Olivier Gaufrès (s243252), Amy Surry (s201902), Lea Skytthe (s203531), Trine Søgaard (s204655)
format:
  html:
    embed-resources: true
editor: visual
---

# First description of the data

## Load libraries

```{r}
#| label: 04_load-libraries
#| message: False
#| warning: False

rm(list=ls())

library("tidyverse")
library("table1")
```

## Load data

```{r}
#| label: 04_load-data
#| message: false

metadata <- read_csv("../data/03_metadata_aug.csv")
```

```{r}
#| label: 04_inspect-data

metadata |> sample_n(5)
```

Making a table of our data

```{r}
#| label: 04_table1

metadata |> 
  mutate(sex = factor(sex),
         disease = factor(disease)) |> 
  table1(x = formula(~ sex + age | disease),
         data = _)
```

## Analysing age, gender, and disease distributions

Lets start with age and sex

```{r}
#| label: 04_age-and-sex-distribution
#| warning: false

metadata |> 
  mutate(age_group = cut(
    x = age,
    breaks = seq(from = 0,
                 to = 100,
                 by = 10))) |> 
  ggplot(mapping = aes(x = age_group, 
                       fill = sex)) +
  geom_histogram(stat = "count", 
                 position = position_dodge(preserve = "single"),
                 color = "black") +
  labs(title = "Age distribution of the patients in our study",
       x = "Age group",
       y = "Count",
       fill = "Gender") +
  theme(plot.title = element_text(hjust = 0.5))
```

Now lets look at disease distribution in our dataset

```{r}
#| label: 04_disease-distribution
#| warning: false

metadata |> 
  ggplot(mapping = aes(x = disease, fill = sex)) +
  geom_histogram(stat = "count",
                 position = position_dodge(preserve = "single"),
                 color = "black") +
  labs(title = "Disease distribution in our study",
       x = "Disease",
       y = "Count",
       fill = "Gender") +
  theme(plot.title = element_text(hjust = 0.5))
```

Do the diseases depend on age?

```{r}
#| label: 04_average-age-for-the-diseases

metadata |> 
  ggplot(mapping = aes(x = disease, y = age, fill = sex)) +
    geom_boxplot() +
  labs(title = "Age distribution for each disease",
       x = "Disease",
       y = "Age",
       fill = "Gender") +
  theme(plot.title = element_text(hjust = 0.5))
```

Lets make a statistical analysis to see if there is any significant difference in the average age of people with cirrhosis disease as it looks like in the above plot.

```{r}
#| label: 04_significance-between-gender

male_cirrhosis <- metadata |> 
  filter(sex == "male", 
         disease == "cirrhosis of liver") |> 
  pull(age)

female_cirrhosis <- metadata |> 
  filter(sex == "female", 
         disease == "cirrhosis of liver") |> 
  pull(age)

p_value <- t.test(male_cirrhosis, female_cirrhosis)$p.value
print(paste("The p-value is:", round(p_value,3)))
```

The p-value is above 0.05 so we can say that there is no significant difference between the two groups.

## Fasting vs postprandial

```{r}
#| label: 04_fasting-and-postprandial-distribution
#| warning: false

metadata |> 
  mutate(age_group = cut(
    x = age,
    breaks = seq(from = 0,
                 to = 100,
                 by = 10))) |> 
  ggplot(mapping = aes(x = age_group, 
                       fill = clinical_information)) +
  geom_histogram(stat = "count", 
                 position = position_dodge(preserve = "single"),
                 color = "black") +
  labs(title = "Age distribution for each study group",
       x = "Age group",
       y = "Count",
       fill = "Study group") +
  theme(plot.title = element_text(hjust = 0.5))
```
